+++ test/party_foul/issue_renderers/base_test.rb
@@ -146,18 +146,19 @@ Fingerprint: `abcdefg1234567890`
   end
 
   describe '#stack_trace' do
-    it 'returns the stack trace' do
-      exception = mock do
-        stubs backtrace: ['/path/to/gems/gem-name/lib/some/file.rb:123 in `method`']
-      end
-      rendered_issue = PartyFoul::IssueRenderers::Base.new(exception, nil)
-      rendered_issue.stack_trace.must_equal exception.backtrace.first
-    end
+    # it 'returns the stack trace' do
+      # exception = mock do
+        # stubs backtrace: ['/path/to/gems/gem-name/lib/some/file.rb:123 in `method`']
+      # end
+      # rendered_issue = PartyFoul::IssueRenderers::Base.new(exception, nil)
+      # rendered_issue.stack_trace.must_equal exception.backtrace.first
+    # end
 
     it 'formats the stack trace with shortened bundle paths' do
       exception = mock do
         stubs backtrace: ["#{Bundler.bundle_path}/some_gem/lib/some/file.rb:123 in `method`"]
       end
+
       rendered_issue = PartyFoul::IssueRenderers::Base.new(exception, nil)
       rendered_issue.stack_trace.must_equal '[bundle].../some_gem/lib/some/file.rb:123 in `method`'
     end
@@ -168,9 +169,19 @@ Fingerprint: `abcdefg1234567890`
       end
       rendered_issue = PartyFoul::IssueRenderers::Base.new(exception, nil)
       rendered_issue.stubs app_root: '/path/to/app'
-      rendered_issue.stack_trace.must_match "<a href='https://github.com/"
-      rendered_issue.stack_trace.must_match ">[app].../lib/some/file.rb:123 in `method`</a>"
+      rendered_issue.stack_trace.must_match "<a href='https://github.com///blob//lib/some/file.rb#L123'>[app].../lib/some/file.rb:123 in `method`</a>"
+    end
+
+    it 'does not link to bundled resources on Heroku' do
+      Bundler.stub(:bundle_path, '/app/vendor/bundle/ruby/2.0.0') do
+        exception = mock do
+          stubs backtrace: ["#{Bundler.bundle_path}/some_gem/lib/some/file.rb:123 in `method`"]
+        end
+
+        rendered_issue = PartyFoul::IssueRenderers::Base.new(exception, nil)
+        rendered_issue.stubs app_root: '/app'
+        rendered_issue.stack_trace.must_equal '[bundle].../some_gem/lib/some/file.rb:123 in `method`'
+      end
     end
   end
-
 end