+++ lib/party_foul/issue_renderers/rails.rb
@@ -4,7 +4,7 @@ class PartyFoul::IssueRenderers::Rails < PartyFoul::IssueRenderers::Rack
   # @return [Hash]
   def params
     parameter_filter = ActionDispatch::Http::ParameterFilter.new(env["action_dispatch.parameter_filter"])
-    parameter_filter.filter(env['action_dispatch.request.path_parameters'])
+    parameter_filter.filter(env['action_dispatch.request.parameters'])
   end
 
   # Rails session hash. Filtered parms are respected.
@@ -30,6 +30,6 @@ class PartyFoul::IssueRenderers::Rails < PartyFoul::IssueRenderers::Rack
   end
 
   def raw_title
-    %{#{env['action_controller.instance'].class}##{env['action_dispatch.request.path_parameters']['action']} (#{exception.class}) "#{exception.message}"}
+    %{#{env['action_controller.instance'].class}##{env['action_dispatch.request.parameters']['action']} (#{exception.class}) "#{exception.message}"}
   end
 end