+++ test/party_foul/issue_renderers/rails_test.rb
@@ -3,7 +3,7 @@ require 'test_helper'
 describe 'Rails Issue Renderer' do
   describe '#params' do
     before do
-      @rendered_issue = PartyFoul::IssueRenderers::Rails.new(nil, {'action_dispatch.parameter_filter' => ['password'], 'action_dispatch.request.path_parameters' => { 'status' => 'ok', 'password' => 'test' }, 'QUERY_STRING' => { 'status' => 'fail' } })
+      @rendered_issue = PartyFoul::IssueRenderers::Rails.new(nil, {'action_dispatch.parameter_filter' => ['password'], 'action_dispatch.request.parameters' => { 'status' => 'ok', 'password' => 'test' }, 'QUERY_STRING' => { 'status' => 'fail' } })
     end
 
     it 'returns ok' do
@@ -53,7 +53,7 @@ describe 'Rails Issue Renderer' do
       controller_instance = mock('Controller')
       controller_instance.stubs(:class).returns('LandingController')
       env = {
-        'action_dispatch.request.path_parameters' => { 'controller' => 'landing', 'action' => 'index' },
+        'action_dispatch.request.parameters' => { 'controller' => 'landing', 'action' => 'index' },
         'action_controller.instance' => controller_instance
       }
       @rendered_issue = PartyFoul::IssueRenderers::Rails.new(@exception, env)