+++ CHANGELOG.md
@@ -5,3 +5,4 @@
    ([Bug #73](https://github.com/dockyard/party_foul/issues/73)) - Dan McClain
 
  * Title prefixing - Javier Martin
+ * Do not link to Heroku bundled gems - Brian Cardarella & Romina Vargas